Don't export p, it's just for us
