+++ spec/installer-spec.js
@@ -2,7 +2,8 @@ import temp from 'temp';
 import jetpack from 'fs-jetpack';
 import sfs from 'fs';
 
-import { createWindowsInstaller, p } from '../src/index.js';
+import { createWindowsInstaller } from '../src/index.js';
+import p from '../src/path-helper';
 
 temp.track();
 
@@ -19,16 +20,16 @@ describe('create-windows-installer task', function() {
 
   it('creates a nuget package and installer', async function() {
     this.timeout(30*1000);
-    
+
     let outputDirectory = temp.mkdirSync('ei-');
-    
+
     let options = {
       appDirectory: p`${__dirname}/fixtures/app`,
       outputDirectory: outputDirectory
     };
-    
+
     await createWindowsInstaller(options);
-    
+
     d(`Verifying assertions on ${outputDirectory}`);
     d(JSON.stringify(sfs.readdirSync(outputDirectory)));
     expect(await jetpack.existsAsync(p`${outputDirectory}/myapp-1.0.0-full.nupkg`)).to.be.ok;
