+++ src/index.js
@@ -7,22 +7,11 @@ import path from 'path';
 import temp from 'temp';
 import jetpack from 'fs-jetpack';
 import fs from 'fs';
+import p from './path-helper';
 
 const d = require('debug')('electron-windows-installer:main');
 
 temp.track();
-
-export function p(strings, ...values) {
-  let newPath = String.raw(strings, ...values);
-  let parts = _.map(newPath.split(/[\\\/]/), (x) => x || '/');
-
-  try {
-    return path.resolve(...parts);
-  } catch(e) {
-    return path.join(...parts);
-  }
-}
-
 async function statNoException(file) {
   try {
     d(file);
